+++ source/platforms/git/gitJSONToGitDSL.ts
@@ -30,8 +30,21 @@ export interface GitJSONToGitDSLConfig {
   /** A promise which will return the string content of a file at a sha */
   getFileContents: (path: string, repo: string | undefined, sha: string) => Promise<string>
   /** A promise which will return the diff string content for a file between shas */
-  getFullDiff: (base: string, head: string) => Promise<string>
+  getFullDiff?: (base: string, head: string) => Promise<string>
+  getFullStructuredDiff?: (base: string, head: string) => Promise<GitStructuredDiff>
 }
+
+export type GitStructuredDiff = {
+  from?: string
+  to?: string
+  chunks: Chunk[]
+}[]
+
+export interface Chunk {
+  changes: Changes
+}
+
+export type Changes = { type: "add" | "del"; content: string }[]
 
 export const gitJSONToGitDSL = (gitJSONRep: GitJSONDSL, config: GitJSONToGitDSLConfig): GitDSL => {
   /**
@@ -133,18 +146,21 @@ export const gitJSONToGitDSL = (gitJSONRep: GitJSONDSL, config: GitJSONToGitDSLC
   const byType = (t: string) => ({ type }: { type: string }) => type === t
   const getContent = ({ content }: { content: string }) => content
 
-  type Changes = { type: string; content: string }[]
-
   /**
    * Gets the git-style diff for a single file.
    *
    * @param filename File path for the diff
    */
   const diffForFile = async (filename: string) => {
-    const diff = await config.getFullDiff(config.baseSHA, config.headSHA)
+    let fileDiffs: GitStructuredDiff
 
-    const fileDiffs: any[] = parseDiff(diff)
-    const structuredDiff = fileDiffs.find((diff: any) => diff.from === filename || diff.to === filename)
+    if (config.getFullStructuredDiff) {
+      fileDiffs = await config.getFullStructuredDiff(config.baseSHA, config.headSHA)
+    } else {
+      const diff = await config.getFullDiff!(config.baseSHA, config.headSHA)
+      fileDiffs = parseDiff(diff)
+    }
+    const structuredDiff = fileDiffs.find(diff => diff.from === filename || diff.to === filename)
 
     if (!structuredDiff) {
       return null