+++ source/runner/jsonToDSL.ts
@@ -6,24 +6,33 @@ import { githubJSONToGitHubDSL } from "../platforms/GitHub"
 import { sentence, href } from "./DangerUtils"
 import { LocalGit } from "../platforms/LocalGit"
 import { GitDSL } from "../dsl/GitDSL"
+import { bitBucketServerGitDSL } from "../platforms/bitbucket_server/BitBucketServerGit"
+import {
+  BitBucketServerAPI,
+  bitbucketServerRepoCredentialsFromEnv,
+} from "../platforms/bitbucket_server/BitBucketServerAPI"
 
 export const jsonToDSL = async (dsl: DangerDSLJSONType): Promise<DangerDSLType> => {
-  const api = githubAPIForDSL(dsl)
-  const platformExists = [dsl.github].some(p => !!p)
-  const github = dsl.github && githubJSONToGitHubDSL(dsl.github, api)
+  const api = apiForDSL(dsl)
+  const platformExists = [dsl.github, dsl.bitbucket_server].some(p => !!p)
+  const github = dsl.github && githubJSONToGitHubDSL(dsl.github, api as GitHubNodeAPI)
+  const bitbucket_server = dsl.bitbucket_server
   // const gitlab = dsl.gitlab && githubJSONToGitLabDSL(dsl.gitlab, api)
 
   let git: GitDSL
   if (!platformExists) {
     const localPlatform = new LocalGit(dsl.settings.cliArgs)
     git = await localPlatform.getPlatformGitRepresentation()
+  } else if (process.env["DANGER_BITBUCKETSERVER_HOST"]) {
+    git = bitBucketServerGitDSL(bitbucket_server!, dsl.git, api as BitBucketServerAPI)
   } else {
-    git = githubJSONToGitDSL(github, dsl.git)
+    git = githubJSONToGitDSL(github!, dsl.git)
   }
 
   return {
     git,
-    github: github,
+    github,
+    bitbucket_server,
     utils: {
       sentence,
       href,
@@ -31,7 +40,11 @@ export const jsonToDSL = async (dsl: DangerDSLJSONType): Promise<DangerDSLType> 
   }
 }
 
-const githubAPIForDSL = (dsl: DangerDSLJSONType) => {
+const apiForDSL = (dsl: DangerDSLJSONType): GitHubNodeAPI | BitBucketServerAPI => {
+  if (process.env["DANGER_BITBUCKETSERVER_HOST"]) {
+    return new BitBucketServerAPI(dsl.bitbucket_server!.metadata, bitbucketServerRepoCredentialsFromEnv(process.env))
+  }
+
   const api = new GitHubNodeAPI({
     host: dsl.settings.github.baseURL,
     headers: {